refactor(aptu-coder): extract path validation helpers into src/validation.rs
